+++ cdk-examples/rust-standalone/src/main.rs
@@ -75,11 +75,21 @@ pub(crate) async fn my_handler(event: Request, ctx: Context) -> Response {
     info!("Congrats! The 1st feature (my-first-feature) is enabled.");
 
     #[cfg(feature = "my-second-feature")]
-    info!("Congrats! The 2nd feature (my-second-feature) is enabled.");
+    {
+        info!("Congrats! The 2nd feature (my-second-feature) is enabled.");
+
+        debug!(
+            "This is a debug message. This line is included to show that we can \
+            conditionally run more than one statement when a feature is enabled."
+        );
+    }
 
     // retrieve an environment variable set at build (compile) time.
     let build_env_var = env!("MY_BUILD_ENV_VAR");
-    debug!("Resolved compile-time value of `MY_BUILD_ENV_VAR`: {}", build_env_var);
+    debug!(
+        "Resolved compile-time value of `MY_BUILD_ENV_VAR`: {}",
+        build_env_var
+    );
 
     // retrieve an environment variable set in the lambda.
     let bucket_name = var("BUCKET_NAME")